Item List improvements
I added a functionality to add item, DOM to handle what happens when mouse is over and out of the button. I added remove button with a function that removes items which were recenty clicked. Separate function gets index of the ckicked element within an array of list elements and stores it in a global varable.